11 endpoints

+++ src/app/api/board/route.js
@@ -1,8 +1,16 @@
 import prisma from "@/lib/prisma";
 
-export async function GET() {
+export async function GET(req) {
   try {
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to view boards",
+      });
+    }
     const boards = await prisma.board.findMany({
+      where: { userId },
       include: {
         columns: { include: { tasks: { include: { subtasks: true } } } },
       },
@@ -23,9 +31,17 @@ export async function POST(req) {
         error: "Please include a name for the board",
       });
     }
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to create boards",
+      });
+    }
     const board = await prisma.board.create({
       data: {
         name,
+        userId,
       },
     });
     return Response.json({ success: true, board });
@@ -34,7 +50,8 @@ export async function POST(req) {
   }
 }
 
-export async function PUT(req) {
+// changes name of board
+export async function PATCH(req) {
   try {
     const data = await req.json();
     const { name, id } = data;
@@ -45,14 +62,26 @@ export async function PUT(req) {
           "Please include the ID of the existing board and what you want to name it",
       });
     }
-
     const board = await prisma.board.findUnique({ where: { id } });
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to edit boards",
+      });
+    }
+    if (userId !== board.userId) {
+      return Response.json({
+        success: false,
+        error: "You are not authorised to edit this board",
+      });
+    }
     if (board) {
-      const newName = await prisma.board.update({
+      const board = await prisma.board.update({
         where: { id },
         data: { name },
       });
-      return Response.json({ success: true, newName });
+      return Response.json({ success: true, board });
     } else {
       return Response.json({
         success: false,
@@ -75,6 +104,19 @@ export async function DELETE(req) {
       });
     }
     const isBoardExisting = await prisma.board.findUnique({ where: { id } });
+    const userId = req.headers.get("userId");
+    if (!userId) {
+      return Response.json({
+        success: false,
+        error: "Please log in to delete boards",
+      });
+    }
+    if (isBoardExisting.userId !== userId) {
+      return Response.json({
+        success: false,
+        error: "Not authorised to delete this board",
+      });
+    }
     if (isBoardExisting) {
       const board = await prisma.board.delete({
         where: { id },
